change datatype in test

+++ tests/test_image_capturing/test_green_detection.py
@@ -2,9 +2,9 @@ from ...image_capturing.green_detection import Blackboard
 import numpy as np
 import cv2
 
-image = np.ones((1080*2, 1920*2, 3), np.uint8) * 255.0
+image = np.ones((1080*2, 1920*2, 3), np.float32) * 255.0
 empty_image = image
-image[250:750, 250:1750] = np.array((0,1,0), np.uint8)*255
+image[250:750, 250:1750] = np.array((0,1,0), np.float32)*255
 
 def test_blackboard_detection():
     blackboard = Blackboard.from_image(image)
